pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/forecasting"]

[project]
name = "datathon2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "holidays>=0.95",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "pmdarima>=2.1.1",
    "pyarrow>=24.0.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
    "catboost>=1.2.7",
    "lightgbm>=4.5.0",
    "prophet>=1.1.5",
    "optuna>=3.6.0",
    "pyyaml>=6.0",
    "torch>=2.4.0",
    "neuralforecast>=1.7.5",
    "lightning>=2.4.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
